+++ tests/test_influxdb_client_3_integration.py
@@ -4,6 +4,7 @@ import os
 import random
 import string
 import time
+import asyncio
 import unittest
 
 import pandas as pd
@@ -16,6 +17,8 @@ from influxdb_client_3 import InfluxDBClient3, write_client_options, WriteOption
 from influxdb_client_3.exceptions import InfluxDBError, InfluxDBPartialWriteError
 from influxdb_client_3.write_client.rest import ApiException
 from tests.util import asyncio_run, lp_to_py_object
+
+running_on_circleci = os.getenv("CIRCLECI") == "true"
 
 
 def random_hex(len=6):
@@ -305,6 +308,29 @@ class TestInfluxDBClient3Integration(unittest.TestCase):
             reader: pyarrow.Table = r_client.query(query, mode="")
             list_results = reader.to_pylist()
             self.assertEqual(data_size, len(list_results))
+
+    @pytest.mark.skipif(running_on_circleci, reason="Skipping this test on CircleCI")
+    def test_multiprocessing_helper(self):
+        org = 'my-org'
+        writer = MultiprocessingWriter(
+            host=self.host,
+            database=self.database,
+            token=self.token,
+            org=org,
+            write_options=WriteOptions(batch_size=1))
+        writer.start()
+        measurement = f'test{random_hex(6)}'.lower()
+        for x in range(1, 10):
+            time.sleep(0.2)
+            writer.write(
+                bucket=self.database,
+                record=f"{measurement},tag=a value=\"number{x}\" {time.time_ns()}"
+            )
+        writer.__del__()
+
+        time.sleep(1)
+        df = self.client.query(f'select * from {measurement}', mode="pandas")
+        self.assertEqual(9, len(df))
 
     test_cert = """-----BEGIN CERTIFICATE-----
 MIIDUzCCAjugAwIBAgIUZB55ULutbc9gy6xLp1BkTQU7siowDQYJKoZIhvcNAQEL
@@ -715,3 +741,15 @@ IdKIRUY6EyIVG+Z/nbuVqUlgnIWOMp0yg4RRC91zHy3Xvykf3Vai25H/jQpa6cbU
             finally:
                 if proxy:
                     proxy.stop()
+
+    def test_call_async_no_exception(self):
+        async def run():
+            await self.client._write_api.post_write_async(
+                "my-org",
+                self.database,
+                "home,room=Sunroom temp=96 1735545600",
+                use_v2_api=False,
+                async_req=True
+            )
+
+        asyncio.run(run())